# api/docs/swagger.yaml
definitions:
  Event.Event:
    properties:
      eventData:
        $ref: '#/definitions/Event.EventData'
      id:
        example: 60b8d295f1d2c916c4d5f1a3
        type: string
    type: object
  Event.EventData:
    properties:
      description:
        example: Описание мероприятия
        type: string
      endDate:
        example: "2025-07-11T21:00:00Z"
        type: string
      name:
        example: Концерт
        type: string
      startDate:
        example: "2025-07-11T18:00:00Z"
        type: string
    type: object
info:
  contact: {}
paths:
  /events:
    get:
      consumes:
      - application/json
      parameters:
      - description: Filter events by date
        enum:
        - past
        - present
        - future
        in: query
        name: filter
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/Event.Event'
            type: array
      summary: Get Events
      tags:
      - Events
    post:
      consumes:
      - application/json
      parameters:
      - description: Event
        in: body
        name: event_data
        required: true
        schema:
          $ref: '#/definitions/Event.EventData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Create Event
      tags:
      - Events
  /events/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "404":
          description: Not Found
          schema: {}
      summary: Delete Event
      tags:
      - Events
    get:
      consumes:
      - application/json
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Event.Event'
      summary: Get Event
      tags:
      - Events
    put:
      consumes:
      - application/json
      parameters:
      - description: EventData
        in: body
        name: event_data
        required: true
        schema:
          $ref: '#/definitions/Event.EventData'
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "404":
          description: Not Found
          schema: {}
      summary: Update Event
      tags:
      - Events
swagger: "2.0"